Fix: crash when fetching website with malformed contents
The problem was that opening tag `<head>` was missing, but `</head>` was present. We were handling this case, but the crash was triggered because there also was `<header>` tag which was falsely identified becase it starts with `<head`. The fix is to search for both variants of the tag (with attributes and without) and also double check to make sure `<head>` is not after `</head>`.